requirements.txt
-i https://pypi.org/simple
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
discord-py==1.3.3
idna-ssl==1.1.0 ; python_version < '3.7'
idna==2.9
multidict==4.7.6
typing-extensions==3.7.4.2 ; python_version < '3.7'
websockets==8.1
yarl==1.4.2
